Move CreateOrUpdateView to utils
