+++ opentech/apply/determinations/views.py
@@ -1,16 +1,13 @@
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import PermissionDenied, ObjectDoesNotExist
-from django.forms import forms
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.urls import reverse_lazy
 from django.utils.decorators import method_decorator
 from django.views.generic import DetailView
-from django.views.generic.detail import SingleObjectTemplateResponseMixin
-from django.views.generic.edit import ProcessFormView, ModelFormMixin
-from django_fsm import can_proceed
 
 from opentech.apply.funds.models import ApplicationSubmission
+from opentech.apply.utils.views import CreateOrUpdateView
 
 from .forms import ConceptDeterminationForm, ProposalDeterminationForm
 from .models import Determination, ACCEPTED, REJECTED
@@ -20,25 +17,6 @@ def get_form_for_stage(submission):
     forms = [ConceptDeterminationForm, ProposalDeterminationForm]
     index = submission.workflow.stages.index(submission.stage)
     return forms[index]
-
-
-class CreateOrUpdateView(SingleObjectTemplateResponseMixin, ModelFormMixin, ProcessFormView):
-
-    def get(self, request, *args, **kwargs):
-        try:
-            self.object = self.get_object()
-        except self.model.DoesNotExist:
-            self.object = None
-
-        return super().get(request, *args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-        try:
-            self.object = self.get_object()
-        except self.model.DoesNotExist:
-            self.object = None
-
-        return super().post(request, *args, **kwargs)
 
 
 class DeterminationCreateOrUpdateView(CreateOrUpdateView):
